align calculating and formatting of dates between components
